feat(refactor): migrate application startup from Startup.cs to Program.cs
- Consolidated application configuration and service registration into Program.cs, enhancing clarity and reducing complexity.
- Removed Startup.cs, integrating its functionality directly into the new minimal hosting model.
- Updated various services and middleware configurations, including logging, CORS, and response compression.
- Improved error handling and IP address capturing for reverse proxy setups.
- Adjusted comments and documentation to reflect the new structure and clarify service registrations.
