+++ Infrastructure/Trackers/Toloka/TolokaParser.cs
@@ -21,249 +21,253 @@ namespace JacRed.Infrastructure.Trackers.Toloka
                 if (string.IsNullOrWhiteSpace(row) || Regex.IsMatch(row, "Збір коштів", RegexOptions.IgnoreCase))
                     continue;
 
-                #region Локальный метод - Match
-                string Match(string pattern, int index = 1)
+                if (!TryParseCreateTime(row, out DateTime createTime))
+                    continue;
+
+                if (!TryParseRowFields(row, out string url, out string title, out string sid, out string pir, out string sizeName))
+                    continue;
+
+                var (name, originalname, relased) = ParseTitleNames(cat, title);
+
+                if (string.IsNullOrWhiteSpace(name))
+                    name = Regex.Split(title, "(\\[|\\/|\\(|\\|)", RegexOptions.IgnoreCase)[0].Trim();
+
+                if (string.IsNullOrWhiteSpace(name))
+                    continue;
+
+                string[] types = GetTypesForCategory(cat);
+                if (types == null)
+                    continue;
+
+                int.TryParse(sid, out int sidNum);
+                int.TryParse(pir, out int pirNum);
+
+                string downloadId = Regex.Match(row, "href=\"download.php\\?id=([0-9]+)\"").Groups[1].Value;
+                if (string.IsNullOrWhiteSpace(downloadId))
+                    continue;
+
+                torrents.Add(new TolokaDetails()
                 {
-                    string res = HttpUtility.HtmlDecode(new Regex(pattern, RegexOptions.IgnoreCase).Match(row).Groups[index].Value.Trim());
-                    res = Regex.Replace(res, "[\n\r\t ]+", " ");
-                    return res.Trim();
+                    trackerName = TrackerName,
+                    types = types,
+                    url = url,
+                    title = title,
+                    sid = sidNum,
+                    pir = pirNum,
+                    sizeName = sizeName,
+                    createTime = createTime,
+                    name = name,
+                    originalname = originalname,
+                    relased = relased,
+                    downloadId = downloadId
+                });
+            }
+
+            return torrents;
+        }
+
+        static string MatchRow(string row, string pattern, int index = 1)
+        {
+            string res = HttpUtility.HtmlDecode(new Regex(pattern, RegexOptions.IgnoreCase).Match(row).Groups[index].Value.Trim());
+            res = Regex.Replace(res, "[\n\r\t ]+", " ");
+            return res.Trim();
+        }
+
+        static bool TryParseCreateTime(string row, out DateTime createTime)
+        {
+            string raw = MatchRow(row, "class=\"postdetails\">([0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2})").Replace("-", ".");
+            if (!DateTime.TryParse(raw, out createTime) || createTime == default)
+            {
+                createTime = default;
+                return false;
+            }
+
+            return true;
+        }
+
+        static bool TryParseRowFields(string row, out string url, out string title, out string sid, out string pir, out string sizeName)
+        {
+            url = MatchRow(row, "<a href=\"(t[0-9]+)\" class=\"topictitle\"");
+            title = MatchRow(row, "class=\"topictitle\">([^<]+)</a>");
+            sid = MatchRow(row, "<span class=\"seedmed\" [^>]+><b>([0-9]+)</b></span>");
+            pir = MatchRow(row, "<span class=\"leechmed\" [^>]+><b>([0-9]+)</b></span>");
+            sizeName = MatchRow(row, "<a href=\"download.php[^\"]+\" [^>]+>([^<]+)</a>").Replace("&nbsp;", " ");
+
+            if (string.IsNullOrWhiteSpace(url) || string.IsNullOrWhiteSpace(title) || string.IsNullOrWhiteSpace(sid) || string.IsNullOrWhiteSpace(pir) || string.IsNullOrWhiteSpace(sizeName) || sizeName == "0 B")
+            {
+                url = title = sid = pir = sizeName = null;
+                return false;
+            }
+
+            url = $"{AppInit.conf.Toloka.host}/{url}";
+            return true;
+        }
+
+        static (string name, string originalname, int relased) ParseTitleNames(string cat, string title)
+        {
+            if (cat is "16" or "96" or "19" or "139" or "12" or "131" or "84" or "42")
+                return ParseMovieTitle(title);
+
+            if (cat is "32" or "173" or "174" or "44" or "230" or "226" or "227" or "228" or "229" or "127" or "124" or "125" or "132")
+                return ParseSerialTitle(title);
+
+            return (null, null, 0);
+        }
+
+        static (string name, string originalname, int relased) ParseMovieTitle(string title)
+        {
+            int relased = 0;
+            string name = null, originalname = null;
+
+            var g = Regex.Match(title, "^([^/\\(\\[]+)/[^/\\(\\[]+/([^/\\(\\[]+) \\(([0-9]{4})(\\)|-)").Groups;
+            if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
+            {
+                name = g[1].Value.Trim();
+                originalname = g[2].Value.Trim();
+                if (int.TryParse(g[3].Value, out int _yer))
+                    relased = _yer;
+            }
+            else
+            {
+                g = Regex.Match(title, "^([^/\\(\\[]+)/([^/\\(\\[]+) \\(([0-9]{4})(\\)|-)").Groups;
+                if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
+                {
+                    name = g[1].Value.Trim();
+                    originalname = g[2].Value.Trim();
+                    if (int.TryParse(g[3].Value, out int _yer))
+                        relased = _yer;
                 }
-                #endregion
-
-                #region Дата создания
-                string _createTime = Match("class=\"postdetails\">([0-9]{4}-[0-9]{2}-[0-9]{2} [0-9]{2}:[0-9]{2})").Replace("-", ".");
-                if (!DateTime.TryParse(_createTime, out DateTime createTime) || createTime == default)
-                    continue;
-                #endregion
-
-                #region Данные раздачи
-                string url = Match("<a href=\"(t[0-9]+)\" class=\"topictitle\"");
-                string title = Match("class=\"topictitle\">([^<]+)</a>");
-
-                string _sid = Match("<span class=\"seedmed\" [^>]+><b>([0-9]+)</b></span>");
-                string _pir = Match("<span class=\"leechmed\" [^>]+><b>([0-9]+)</b></span>");
-                string sizeName = Match("<a href=\"download.php[^\"]+\" [^>]+>([^<]+)</a>").Replace("&nbsp;", " ");
-
-                if (string.IsNullOrWhiteSpace(url) || string.IsNullOrWhiteSpace(title) || string.IsNullOrWhiteSpace(_sid) || string.IsNullOrWhiteSpace(_pir) || string.IsNullOrWhiteSpace(sizeName) || sizeName == "0 B")
-                    continue;
-
-                url = $"{AppInit.conf.Toloka.host}/{url}";
-                #endregion
-
-                #region Парсим раздачи
-                int relased = 0;
-                string name = null, originalname = null;
-
-                if (cat is "16" or "96" or "19" or "139" or "12" or "131" or "84" or "42")
+                else
                 {
-                    #region Фильмы
-                    // Незворотність / Irréversible / Irreversible (2002) AVC Ukr/Fre | Sub Eng
-                    var g = Regex.Match(title, "^([^/\\(\\[]+)/[^/\\(\\[]+/([^/\\(\\[]+) \\(([0-9]{4})(\\)|-)").Groups;
-                    if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
+                    g = Regex.Match(title, "^([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
+                    if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value))
                     {
-                        name = g[1].Value.Trim();
-                        originalname = g[2].Value.Trim();
-
-                        if (int.TryParse(g[3].Value, out int _yer))
+                        name = g[1].Value;
+                        if (int.TryParse(g[2].Value, out int _yer))
                             relased = _yer;
                     }
                     else
                     {
-                        // Мій рік у Нью-Йорку / My Salinger Year (2020) Ukr/Eng
-                        g = Regex.Match(title, "^([^/\\(\\[]+)/([^/\\(\\[]+) \\(([0-9]{4})(\\)|-)").Groups;
+                        g = Regex.Match(title, "^([^/\\(\\[]+) \\(([0-9]{4})(\\)|-)").Groups;
+                        if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value))
+                        {
+                            name = g[1].Value;
+                            if (int.TryParse(g[2].Value, out int _yer))
+                                relased = _yer;
+                        }
+                    }
+                }
+            }
+
+            return (name, originalname, relased);
+        }
+
+        static (string name, string originalname, int relased) ParseSerialTitle(string title)
+        {
+            int relased = 0;
+            string name = null, originalname = null;
+
+            var g = Regex.Match(title, "^([^/\\(\\[]+) \\([^\\)]+\\) \\([^\\)]+\\) ?/([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
+            if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
+            {
+                name = g[1].Value.Trim();
+                originalname = g[2].Value.Trim();
+                if (int.TryParse(g[3].Value, out int _yer))
+                    relased = _yer;
+            }
+            else
+            {
+                g = Regex.Match(title, "^([^/\\(\\[]+) \\([^\\)]+\\) ?/([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
+                if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
+                {
+                    name = g[1].Value.Trim();
+                    originalname = g[2].Value.Trim();
+                    if (int.TryParse(g[3].Value, out int _yer))
+                        relased = _yer;
+                }
+                else
+                {
+                    g = Regex.Match(title, "^([^/\\(\\[]+) (\\(|\\[)[^\\)\\]]+(\\)|\\]) ?/([^/\\(\\[]+) \\(([0-9]{4})(\\)|-)").Groups;
+                    if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[4].Value) && !string.IsNullOrWhiteSpace(g[5].Value))
+                    {
+                        name = g[1].Value.Trim();
+                        originalname = g[4].Value.Trim();
+                        if (int.TryParse(g[5].Value, out int _yer))
+                            relased = _yer;
+                    }
+                    else
+                    {
+                        g = Regex.Match(title, "^([^/\\(\\[]+)/([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
                         if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
                         {
                             name = g[1].Value.Trim();
                             originalname = g[2].Value.Trim();
-
                             if (int.TryParse(g[3].Value, out int _yer))
                                 relased = _yer;
                         }
                         else
                         {
-                            // Хроніка надій та ілюзій. Дзеркало історії. (83 серії) (2001-2003) PDTVRip
-                            g = Regex.Match(title, "^([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
-                            if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value))
+                            g = Regex.Match(title, "^([^/\\(\\[]+)/[^/\\(\\[]+/([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
+                            if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
                             {
-                                name = g[1].Value;
-
-                                if (int.TryParse(g[2].Value, out int _yer))
+                                name = g[1].Value.Trim();
+                                originalname = g[2].Value.Trim();
+                                if (int.TryParse(g[3].Value, out int _yer))
                                     relased = _yer;
                             }
                             else
                             {
-                                // Берестечко. Битва за Україну (2015-2016) DVDRip-AVC
-                                g = Regex.Match(title, "^([^/\\(\\[]+) \\(([0-9]{4})(\\)|-)").Groups;
+                                g = Regex.Match(title, "^([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
                                 if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value))
                                 {
-                                    name = g[1].Value;
-
+                                    name = g[1].Value.Trim();
                                     if (int.TryParse(g[2].Value, out int _yer))
                                         relased = _yer;
                                 }
                             }
                         }
                     }
-                    #endregion
-                }
-                else if (cat is "32" or "173" or "174" or "44" or "230" or "226" or "227" or "228" or "229" or "127" or "124" or "125" or "132")
-                {
-                    #region Сериалы
-                    // Атака титанів (Attack on Titan) (Сезон 1) / Shingeki no Kyojin (Season 1) (2013) BDRip 720р
-                    var g = Regex.Match(title, "^([^/\\(\\[]+) \\([^\\)]+\\) \\([^\\)]+\\) ?/([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
-                    if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
-                    {
-                        name = g[1].Value.Trim();
-                        originalname = g[2].Value.Trim();
-
-                        if (int.TryParse(g[3].Value, out int _yer))
-                            relased = _yer;
-                    }
-                    else
-                    {
-                        // Дім з прислугою (Сезон 2, серії 1-8) / Servant (Season 2, episodes 1-8) (2021) WEB-DLRip-AVC Ukr/Eng
-                        g = Regex.Match(title, "^([^/\\(\\[]+) \\([^\\)]+\\) ?/([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
-                        if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
-                        {
-                            name = g[1].Value.Trim();
-                            originalname = g[2].Value.Trim();
-
-                            if (int.TryParse(g[3].Value, out int _yer))
-                                relased = _yer;
-                        }
-                        else
-                        {
-                            // Детективне агентство прекрасних хлопчиків (08 з 12) / Bishounen Tanteidan (2021) BDRip 1080p Ukr/Jap | Ukr Sub
-                            g = Regex.Match(title, "^([^/\\(\\[]+) (\\(|\\[)[^\\)\\]]+(\\)|\\]) ?/([^/\\(\\[]+) \\(([0-9]{4})(\\)|-)").Groups;
-                            if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[4].Value) && !string.IsNullOrWhiteSpace(g[5].Value))
-                            {
-                                name = g[1].Value.Trim();
-                                originalname = g[4].Value.Trim();
-
-                                if (int.TryParse(g[5].Value, out int _yer))
-                                    relased = _yer;
-                            }
-                            else
-                            {
-                                // Яйця Дракона / Dragon Ball (01-31 з 153) (1986-1989) BDRip 1080p H.265
-                                // Томо — дівчина! / Tomo-chan wa Onnanoko! (Сезон 1, серії 01-02 з 13) (2023) WEBDL 1080p H.265 Ukr/Jap | sub Ukr
-                                g = Regex.Match(title, "^([^/\\(\\[]+)/([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
-                                if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
-                                {
-                                    name = g[1].Value.Trim();
-                                    originalname = g[2].Value.Trim();
-
-                                    if (int.TryParse(g[3].Value, out int _yer))
-                                        relased = _yer;
-                                }
-                                else
-                                {
-                                    // Людина-бензопила / チェンソーマン /Chainsaw Man (сезон 1, серії 8 з 12) (2022) WEBRip 1080p
-                                    g = Regex.Match(title, "^([^/\\(\\[]+)/[^/\\(\\[]+/([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
-                                    if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value) && !string.IsNullOrWhiteSpace(g[3].Value))
-                                    {
-                                        name = g[1].Value.Trim();
-                                        originalname = g[2].Value.Trim();
-
-                                        if (int.TryParse(g[3].Value, out int _yer))
-                                            relased = _yer;
-                                    }
-                                    else
-                                    {
-                                        // МастерШеф. 10 сезон (1-18 епізоди) (2020) IPTVRip 400p
-                                        g = Regex.Match(title, "^([^/\\(\\[]+) \\([^\\)]+\\) \\(([0-9]{4})(\\)|-)").Groups;
-                                        if (!string.IsNullOrWhiteSpace(g[1].Value) && !string.IsNullOrWhiteSpace(g[2].Value))
-                                        {
-                                            name = g[1].Value.Trim();
-
-                                            if (int.TryParse(g[2].Value, out int _yer))
-                                                relased = _yer;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    #endregion
-                }
-                #endregion
-
-                if (string.IsNullOrWhiteSpace(name))
-                    name = Regex.Split(title, "(\\[|\\/|\\(|\\|)", RegexOptions.IgnoreCase)[0].Trim();
-
-                if (!string.IsNullOrWhiteSpace(name))
-                {
-                    #region types
-                    string[] types = null;
-                    switch (cat)
-                    {
-                        case "16":
-                        case "96":
-                        case "42":
-                            types = new string[] { "movie" };
-                            break;
-                        case "19":
-                        case "139":
-                        case "84":
-                            types = new string[] { "multfilm" };
-                            break;
-                        case "32":
-                        case "173":
-                        case "124":
-                            types = new string[] { "serial" };
-                            break;
-                        case "174":
-                        case "44":
-                        case "125":
-                            types = new string[] { "multserial" };
-                            break;
-                        case "226":
-                        case "227":
-                        case "228":
-                        case "229":
-                        case "230":
-                        case "12":
-                        case "131":
-                            types = new string[] { "docuserial", "documovie" };
-                            break;
-                        case "127":
-                            types = new string[] { "anime" };
-                            break;
-                        case "132":
-                            types = new string[] { "tvshow" };
-                            break;
-                    }
-
-                    if (types == null)
-                        continue;
-                    #endregion
-
-                    int.TryParse(_sid, out int sid);
-                    int.TryParse(_pir, out int pir);
-
-                    string downloadId = Regex.Match(row, "href=\"download.php\\?id=([0-9]+)\"").Groups[1].Value;
-                    if (string.IsNullOrWhiteSpace(downloadId))
-                        continue;
-
-                    torrents.Add(new TolokaDetails()
-                    {
-                        trackerName = TrackerName,
-                        types = types,
-                        url = url,
-                        title = title,
-                        sid = sid,
-                        pir = pir,
-                        sizeName = sizeName,
-                        createTime = createTime,
-                        name = name,
-                        originalname = originalname,
-                        relased = relased,
-                        downloadId = downloadId
-                    });
                 }
             }
 
-            return torrents;
+            return (name, originalname, relased);
+        }
+
+        static string[] GetTypesForCategory(string cat)
+        {
+            switch (cat)
+            {
+                case "16":
+                case "96":
+                case "42":
+                    return new string[] { "movie" };
+                case "19":
+                case "139":
+                case "84":
+                    return new string[] { "multfilm" };
+                case "32":
+                case "173":
+                case "124":
+                    return new string[] { "serial" };
+                case "174":
+                case "44":
+                case "125":
+                    return new string[] { "multserial" };
+                case "226":
+                case "227":
+                case "228":
+                case "229":
+                case "230":
+                case "12":
+                case "131":
+                    return new string[] { "docuserial", "documovie" };
+                case "127":
+                    return new string[] { "anime" };
+                case "132":
+                    return new string[] { "tvshow" };
+                default:
+                    return null;
+            }
         }
     }
 }
